extract state from component to state, utils, simplify types and custom render func.
