+++ examples/cookbook/jotai/test-utils.tsx
@@ -1,13 +1,22 @@
 import * as React from 'react';
-import {render} from '@testing-library/react-native';
-import {useHydrateAtoms} from "jotai/utils";
-import {IHydrateAtomsProps, InitialValues, IRenderWithAtomsOptions} from "./types";
+import { render } from '@testing-library/react-native';
+import { useHydrateAtoms } from "jotai/utils";
+import { HydrateAtomsWrapperProps, RenderWithAtomsOptions } from "./types";
 
+/**
+ * A wrapper component that hydrates Jotai atoms with initial values.
+ *
+ * @template T - The type of the initial values for the atoms.
+ * @param initialValues - The initial values for the Jotai atoms.
+ * @param children - The child components to render.
+ * @returns The rendered children.
+
+ */
 function HydrateAtomsWrapper<T>({
-                                initialValues,
-                                children,
-                              }: IHydrateAtomsProps<T>) {
-  useHydrateAtoms(initialValues as unknown as InitialValues);
+                                  initialValues,
+                                  children,
+                                }: HydrateAtomsWrapperProps<T>) {
+  useHydrateAtoms(initialValues);
   return children;
 }
 
@@ -21,15 +30,13 @@ function HydrateAtomsWrapper<T>({
  */
 export const renderWithAtoms = <T, >(
   component: React.ReactElement,
-  options: IRenderWithAtomsOptions<T>,
+  options: RenderWithAtomsOptions<T>,
 ) => {
-  const {initialValues} = options;
-  return render(component, {
-    wrapper: ({children}: { children: React.JSX.Element }) => (
-      <HydrateAtomsWrapper initialValues={initialValues}>
-        {children}
-      </HydrateAtomsWrapper>
-    ),
-    ...options,
-  });
+  const {initialValues, ...rest} = options;
+
+  const ui = <HydrateAtomsWrapper initialValues={initialValues}>
+    {component}
+  </HydrateAtomsWrapper>;
+
+  return render(ui, {...rest});
 };
