+++ examples/cookbook/jotai/utils.ts
@@ -1,0 +1,1 @@
+export const generateRandomId = () => Math.random().toString(36).slice(2, 11);
